refactor: Make convert date method reusable
