+++ client/src/Pages/Home/Components/Featured/Components/ArticleCard.js
@@ -1,14 +1,6 @@
 import { Link } from 'react-router-dom'
 import './ArticleCard.css'
-
-const convertDate = (dateStr) => {
-    const date = new Date(dateStr)
-    const year = date.getFullYear()
-    const month = String(date.getMonth()+1).padStart(2,'0')
-    const day = String(date.getDate()).padStart(2,'0')
-
-    return `${year}.${month}.${day}`
-}
+import convertDate from 'Utility/convertDate'
 
 export default function ArticleCard({article}) {
   return (
