+++ client/src/Utility/convertDate.js
@@ -1,0 +1,8 @@
+export default function convertDate(dateStr) {
+    const date = new Date(dateStr)
+    const year = date.getFullYear()
+    const month = String(date.getMonth()+1).padStart(2,'0')
+    const day = String(date.getDate()).padStart(2,'0')
+
+    return `${year}.${month}.${day}`
+}
